requirements.txt
pystray==0.19.5
Pillow==10.1.0
psutil==5.9.6
sqlalchemy==2.0.23
alembic==1.12.1
pywin32==306